Big upgrade
This push includes a significant upgrade to the project, featuring the implementation of a server-side built with PostgreSQL and Express. As part of enhancing security measures, the .env file has been removed from the project to safeguard personal usage.

Notable additions include the client-side implementation of an Admin page, which facilitates the management of game levels, card types, themes, and individual cards, thereby expanding the game's variety. Additionally, an Authentication page and registration logic have been incorporated, enabling users to sign in and access the game.

Furthermore, the game logic has been reworked to align with the server-side architecture, resulting in a seamless experience.

Overall, this push represents a substantial advancement of the project, offering improved security, expanded functionality, and enhanced game logic.